Build:WIP:From now on, I'll stop working on i18n files (.po, .pot, .mo) and concentrate on my specific json format. We'll look at the i18n catalog and so on later.
